Rename SplitFiles to SplitFileReferences
