+++ java/core/src/test/java/sleeper/core/statestore/SplitFileReferencesTest.java
@@ -14,48 +14,44 @@
  * limitations under the License.
  */
 
-package sleeper.statestore.s3;
+package sleeper.core.statestore;
 
-import org.apache.hadoop.conf.Configuration;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.partition.PartitionsBuilder;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.LongType;
-import sleeper.core.statestore.FileReference;
-import sleeper.core.statestore.FileReferenceFactory;
-import sleeper.core.statestore.SplitFileReference;
-import sleeper.core.statestore.SplitFiles;
-import sleeper.core.statestore.StateStore;
-import sleeper.core.statestore.StateStoreException;
 
 import java.time.Instant;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.properties.table.TablePropertiesTestHelper.createTestTableProperties;
 import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
+import static sleeper.core.statestore.inmemory.StateStoreTestHelper.inMemoryStateStoreWithNoPartitions;
 
-public class S3SplitFilesIT extends S3StateStoreTestBase {
+public class SplitFileReferencesTest {
     private static final Instant DEFAULT_UPDATE_TIME = Instant.parse("2023-10-04T14:08:00Z");
     private final Schema schema = schemaWithKey("key", new LongType());
     private final PartitionsBuilder partitions = new PartitionsBuilder(schema).singlePartition("root");
     private FileReferenceFactory factory = FileReferenceFactory.fromUpdatedAt(partitions.buildTree(), DEFAULT_UPDATE_TIME);
+    private final StateStore store = inMemoryStateStoreWithNoPartitions();
+
+    @BeforeEach
+    void setUp() {
+        store.fixTime(DEFAULT_UPDATE_TIME);
+    }
 
     @Test
     void shouldSplitOneFileInNonLeafPartition() throws Exception {
         // Given
-        StateStore store = createStateStore(schema);
-        store.fixTime(DEFAULT_UPDATE_TIME);
-        splitPartition(store, "root", "L", "R", 5L);
+        splitPartition("root", "L", "R", 5L);
         FileReference file = factory.rootFile("to-split.parquet", 100L);
-
         store.addFile(file);
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -67,15 +63,13 @@ public class S3SplitFilesIT extends S3StateStoreTestBase {
     @Test
     void shouldNotSplitOneFileInLeafPartition() throws Exception {
         // Given
-        StateStore store = createStateStore(schema);
-        store.fixTime(DEFAULT_UPDATE_TIME);
-        splitPartition(store, "root", "L", "R", 5L);
+        splitPartition("root", "L", "R", 5L);
         FileReference file = factory.partitionFile("L", "already-split.parquet", 100L);
         store.addFile(file);
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -85,18 +79,16 @@ public class S3SplitFilesIT extends S3StateStoreTestBase {
     @Test
     void shouldSplitTwoFilesInDifferentPartitions() throws Exception {
         // Given
-        StateStore store = createStateStore(schema);
-        store.fixTime(DEFAULT_UPDATE_TIME);
-        splitPartition(store, "root", "L", "R", 5L);
-        splitPartition(store, "L", "LL", "LR", 2L);
-        splitPartition(store, "R", "RL", "RR", 7L);
+        splitPartition("root", "L", "R", 5L);
+        splitPartition("L", "LL", "LR", 2L);
+        splitPartition("R", "RL", "RR", 7L);
         FileReference file1 = factory.partitionFile("L", "file1.parquet", 100L);
         FileReference file2 = factory.partitionFile("R", "file2.parquet", 100L);
         store.addFiles(List.of(file1, file2));
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -110,17 +102,15 @@ public class S3SplitFilesIT extends S3StateStoreTestBase {
     @Test
     void shouldOnlyPerformOneLevelOfSplits() throws Exception {
         // Given
-        StateStore store = createStateStore(schema);
-        store.fixTime(DEFAULT_UPDATE_TIME);
-        splitPartition(store, "root", "L", "R", 5L);
-        splitPartition(store, "L", "LL", "LR", 2L);
-        splitPartition(store, "R", "RL", "RR", 7L);
+        splitPartition("root", "L", "R", 5L);
+        splitPartition("L", "LL", "LR", 2L);
+        splitPartition("R", "RL", "RR", 7L);
         FileReference file = factory.rootFile("file.parquet", 100L);
         store.addFile(file);
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -129,18 +119,7 @@ public class S3SplitFilesIT extends S3StateStoreTestBase {
                         splitFile(file, "R"));
     }
 
-    private S3StateStore createStateStore(Schema schema) {
-        TableProperties tableProperties = createTestTableProperties(instanceProperties, schema);
-        S3StateStore stateStore = new S3StateStore(instanceProperties, tableProperties, dynamoDBClient, new Configuration());
-        try {
-            stateStore.initialise();
-            return stateStore;
-        } catch (StateStoreException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void splitPartition(StateStore store, String parentId, String leftId, String rightId, long splitPoint) throws StateStoreException {
+    private void splitPartition(String parentId, String leftId, String rightId, long splitPoint) throws StateStoreException {
         partitions.splitToNewChildren(parentId, leftId, rightId, splitPoint)
                 .applySplit(store, parentId);
         factory = FileReferenceFactory.fromUpdatedAt(partitions.buildTree(), DEFAULT_UPDATE_TIME);
