+++ java/statestore/src/test/java/sleeper/statestore/s3/S3SplitFileReferencesIT.java
@@ -14,44 +14,48 @@
  * limitations under the License.
  */
 
-package sleeper.core.statestore;
+package sleeper.statestore.s3;
 
-import org.junit.jupiter.api.BeforeEach;
+import org.apache.hadoop.conf.Configuration;
 import org.junit.jupiter.api.Test;
 
+import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.partition.PartitionsBuilder;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.LongType;
+import sleeper.core.statestore.FileReference;
+import sleeper.core.statestore.FileReferenceFactory;
+import sleeper.core.statestore.SplitFileReference;
+import sleeper.core.statestore.SplitFileReferences;
+import sleeper.core.statestore.StateStore;
+import sleeper.core.statestore.StateStoreException;
 
 import java.time.Instant;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static sleeper.configuration.properties.table.TablePropertiesTestHelper.createTestTableProperties;
 import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
-import static sleeper.core.statestore.inmemory.StateStoreTestHelper.inMemoryStateStoreWithNoPartitions;
 
-public class SplitFilesTest {
+public class S3SplitFileReferencesIT extends S3StateStoreTestBase {
     private static final Instant DEFAULT_UPDATE_TIME = Instant.parse("2023-10-04T14:08:00Z");
     private final Schema schema = schemaWithKey("key", new LongType());
     private final PartitionsBuilder partitions = new PartitionsBuilder(schema).singlePartition("root");
     private FileReferenceFactory factory = FileReferenceFactory.fromUpdatedAt(partitions.buildTree(), DEFAULT_UPDATE_TIME);
-    private final StateStore store = inMemoryStateStoreWithNoPartitions();
-
-    @BeforeEach
-    void setUp() {
-        store.fixTime(DEFAULT_UPDATE_TIME);
-    }
 
     @Test
     void shouldSplitOneFileInNonLeafPartition() throws Exception {
         // Given
-        splitPartition("root", "L", "R", 5L);
+        StateStore store = createStateStore(schema);
+        store.fixTime(DEFAULT_UPDATE_TIME);
+        splitPartition(store, "root", "L", "R", 5L);
         FileReference file = factory.rootFile("to-split.parquet", 100L);
+
         store.addFile(file);
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -63,13 +67,15 @@ public class SplitFilesTest {
     @Test
     void shouldNotSplitOneFileInLeafPartition() throws Exception {
         // Given
-        splitPartition("root", "L", "R", 5L);
+        StateStore store = createStateStore(schema);
+        store.fixTime(DEFAULT_UPDATE_TIME);
+        splitPartition(store, "root", "L", "R", 5L);
         FileReference file = factory.partitionFile("L", "already-split.parquet", 100L);
         store.addFile(file);
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -79,16 +85,18 @@ public class SplitFilesTest {
     @Test
     void shouldSplitTwoFilesInDifferentPartitions() throws Exception {
         // Given
-        splitPartition("root", "L", "R", 5L);
-        splitPartition("L", "LL", "LR", 2L);
-        splitPartition("R", "RL", "RR", 7L);
+        StateStore store = createStateStore(schema);
+        store.fixTime(DEFAULT_UPDATE_TIME);
+        splitPartition(store, "root", "L", "R", 5L);
+        splitPartition(store, "L", "LL", "LR", 2L);
+        splitPartition(store, "R", "RL", "RR", 7L);
         FileReference file1 = factory.partitionFile("L", "file1.parquet", 100L);
         FileReference file2 = factory.partitionFile("R", "file2.parquet", 100L);
         store.addFiles(List.of(file1, file2));
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -102,15 +110,17 @@ public class SplitFilesTest {
     @Test
     void shouldOnlyPerformOneLevelOfSplits() throws Exception {
         // Given
-        splitPartition("root", "L", "R", 5L);
-        splitPartition("L", "LL", "LR", 2L);
-        splitPartition("R", "RL", "RR", 7L);
+        StateStore store = createStateStore(schema);
+        store.fixTime(DEFAULT_UPDATE_TIME);
+        splitPartition(store, "root", "L", "R", 5L);
+        splitPartition(store, "L", "LL", "LR", 2L);
+        splitPartition(store, "R", "RL", "RR", 7L);
         FileReference file = factory.rootFile("file.parquet", 100L);
         store.addFile(file);
 
         // When
-        SplitFiles splitFiles = SplitFiles.from(store);
-        splitFiles.split();
+        SplitFileReferences splitFileReferences = SplitFileReferences.from(store);
+        splitFileReferences.split();
 
         // Then
         assertThat(store.getActiveFiles())
@@ -119,7 +129,18 @@ public class SplitFilesTest {
                         splitFile(file, "R"));
     }
 
-    private void splitPartition(String parentId, String leftId, String rightId, long splitPoint) throws StateStoreException {
+    private S3StateStore createStateStore(Schema schema) {
+        TableProperties tableProperties = createTestTableProperties(instanceProperties, schema);
+        S3StateStore stateStore = new S3StateStore(instanceProperties, tableProperties, dynamoDBClient, new Configuration());
+        try {
+            stateStore.initialise();
+            return stateStore;
+        } catch (StateStoreException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void splitPartition(StateStore store, String parentId, String leftId, String rightId, long splitPoint) throws StateStoreException {
         partitions.splitToNewChildren(parentId, leftId, rightId, splitPoint)
                 .applySplit(store, parentId);
         factory = FileReferenceFactory.fromUpdatedAt(partitions.buildTree(), DEFAULT_UPDATE_TIME);
